+++ src/app/auth_service/urls.py
@@ -1,10 +1,11 @@
-from fastapi import APIRouter, status
+from fastapi import APIRouter, UploadFile, status
 
 from app.auth_service.schemas import TokenSchema, UserSchema
 from app.auth_service.views import (
     check_token_dependency,
     login_view,
     register_view,
+    verify_view,
 )
 
 router = APIRouter(tags=['auth_service'])
@@ -37,3 +38,13 @@ async def login(user_in: UserSchema) -> TokenSchema:
 async def check_token(user_id: int) -> None:
     """Проверка токена. Возвращает имя пользователя."""
     await check_token_dependency(user_id)
+
+
+@router.post(
+    '/verify/',
+    status_code=status.HTTP_201_CREATED,
+)
+async def verify(user_photo: UploadFile, user_id: int) -> dict:
+    """Подтверждение пользователя."""
+    await check_token_dependency(user_id)
+    return await verify_view(user_photo, user_id)