+++ src/app/settings.py
@@ -2,7 +2,7 @@ class Settings():
     """Класс настроек."""
 
     auth_service_url: str = 'http://host.docker.internal:8001'
-    transactions_service_url: str = 'http://host.docker.internal:8002/transactions'  # noqa: E501
+    transactions_service_url: str = 'http://host.docker.internal:8002'
     face_verification_service_url: str = 'http://host.docker.internal:8003'